moved convenience functions to basechart.js and fixed an embarrassing spelling error
